+++ ModelsForm.cs
@@ -38,13 +38,21 @@ namespace NissanManufacturingPlanning
 
         private void button3_Click(object sender, EventArgs e)
         {
-            string name = dgvModel.Rows[dgvModel.SelectedCells[0].RowIndex].Cells["ModelName"].Value.ToString();
-            string index = dgvModel.Rows[dgvModel.SelectedCells[0].RowIndex].Cells["ModelID"].Value.ToString();
-            if (DialogResult.Yes == MessageBox.Show("Do you want to Delete " + name + "?", "Confirmation", MessageBoxButtons.YesNo, MessageBoxIcon.Warning))
+            if (dgvModel.Rows.Count == 0)
             {
-                new MainForm().SqlDelete("DELETE FROM VehicleModel WHERE ModelID = " + index);
+                MessageBox.Show("No Models to Remove");
             }
-            new MainForm().QuerySelectAll("VehicleModel", dgvModel);
+            else
+            {
+                string name = dgvModel.Rows[dgvModel.SelectedCells[0].RowIndex].Cells["ModelName"].Value.ToString();
+                string index = dgvModel.Rows[dgvModel.SelectedCells[0].RowIndex].Cells["ModelID"].Value.ToString();
+                if (DialogResult.Yes == MessageBox.Show("Do you want to Delete " + name + "?", "Confirmation", MessageBoxButtons.YesNo, MessageBoxIcon.Warning))
+                {
+                    new MainForm().SqlDelete("DELETE FROM VehicleModel WHERE ModelID = " + index);
+                }
+                new MainForm().QuerySelectAll("VehicleModel", dgvModel);
+            }
+            
         }
     }
 }
