+++ PlantsForm.cs
@@ -36,34 +36,50 @@ namespace NissanManufacturingPlanning
 
         private void button2_Click(object sender, EventArgs e)
         {
-            PlantForm plant = new PlantForm();
-            plant.btnAdd.Text = "Update Plant";
-            plant.tbxName.Text = dgvPlants.Rows[dgvPlants.SelectedCells[0].RowIndex].Cells["Name"].Value.ToString(); 
-            plant.tbxCountry.Text = dgvPlants.Rows[dgvPlants.SelectedCells[0].RowIndex].Cells["Country"].Value.ToString(); 
-            plant.tbxAddress.Text = dgvPlants.Rows[dgvPlants.SelectedCells[0].RowIndex].Cells["Address"].Value.ToString(); 
-            plant.numDuration.Value = Convert.ToInt32(dgvPlants.Rows[dgvPlants.SelectedCells[0].RowIndex].Cells["ShiftDuration"].Value.ToString()) / 60; 
-            plant.numRate.Value = Convert.ToInt32(dgvPlants.Rows[dgvPlants.SelectedCells[0].RowIndex].Cells["Rate"].Value.ToString()); 
-            plant.numUptime.Value = Convert.ToInt32(dgvPlants.Rows[dgvPlants.SelectedCells[0].RowIndex].Cells["Uptime"].Value.ToString());
-            plant.ShowDialog();
+            if (dgvPlants.Rows.Count == 0)//Null/Empty Datagridview checking on Update
+            {
+                MessageBox.Show("No Plants to Update");
+            }else
+            {
+                PlantForm plant = new PlantForm();
+                plant.btnAdd.Text = "Update Plant";
+                plant.tbxName.Text = dgvPlants.Rows[dgvPlants.SelectedCells[0].RowIndex].Cells["Name"].Value.ToString();
+                plant.tbxCountry.Text = dgvPlants.Rows[dgvPlants.SelectedCells[0].RowIndex].Cells["Country"].Value.ToString();
+                plant.tbxAddress.Text = dgvPlants.Rows[dgvPlants.SelectedCells[0].RowIndex].Cells["Address"].Value.ToString();
+                plant.numDuration.Value = Convert.ToInt32(dgvPlants.Rows[dgvPlants.SelectedCells[0].RowIndex].Cells["ShiftDuration"].Value.ToString()) / 60;
+                plant.numRate.Value = Convert.ToInt32(dgvPlants.Rows[dgvPlants.SelectedCells[0].RowIndex].Cells["Rate"].Value.ToString());
+                plant.numUptime.Value = Convert.ToInt32(dgvPlants.Rows[dgvPlants.SelectedCells[0].RowIndex].Cells["Uptime"].Value.ToString());
+                plant.ShowDialog();
 
-            string index = dgvPlants.Rows[dgvPlants.SelectedCells[0].RowIndex].Cells["PlantID"].Value.ToString();
-            
-            if (plant.fPlant != null)
-            { 
-                new MainForm().SqlUpdate("UPDATE Plant SET Name = '" + plant.fPlant.getName() + "', Country = '" + plant.fPlant.getCountryId() + "', Address = '" + plant.fPlant.getAddress() + "', Rate = " + plant.fPlant.getRate().ToString() + ", Uptime = " + plant.fPlant.getAvailablity().ToString() + ", ShiftDuration = " + plant.fPlant.getShiftDuration().ToString() + " WHERE PlantID = '" + index + "'");
-                new MainForm().QuerySelectAll("Plant", dgvPlants);
+                string index = dgvPlants.Rows[dgvPlants.SelectedCells[0].RowIndex].Cells["PlantID"].Value.ToString();
+
+                if (plant.fPlant != null)
+                {
+                    new MainForm().SqlUpdate("UPDATE Plant SET Name = '" + plant.fPlant.getName() + "', Country = '" + plant.fPlant.getCountryId() + "', Address = '" + plant.fPlant.getAddress() + "', Rate = " + plant.fPlant.getRate().ToString() + ", Uptime = " + plant.fPlant.getAvailablity().ToString() + ", ShiftDuration = " + plant.fPlant.getShiftDuration().ToString() + " WHERE PlantID = '" + index + "'");
+                    new MainForm().QuerySelectAll("Plant", dgvPlants);
+                }
             }
+            
         }
 
         private void button3_Click(object sender, EventArgs e)
         {
-            string name = dgvPlants.Rows[dgvPlants.SelectedCells[0].RowIndex].Cells["Name"].Value.ToString();
-            string index = dgvPlants.Rows[dgvPlants.SelectedCells[0].RowIndex].Cells["PlantID"].Value.ToString();
-            if (DialogResult.Yes == MessageBox.Show("Do you want to Delete "+name+"?","Confirmation", MessageBoxButtons.YesNo, MessageBoxIcon.Warning))
+            
+            if (dgvPlants.Rows.Count == 0)//Null checking on Remove
             {
-                new MainForm().SqlDelete("DELETE FROM Plant WHERE PlantID = "+index);
+                MessageBox.Show("No Plants to Remove");
             }
-            new MainForm().QuerySelectAll("Plant", dgvPlants);
+            else
+            {
+                string name = dgvPlants.Rows[dgvPlants.SelectedCells[0].RowIndex].Cells["Name"].Value.ToString();
+                string index = dgvPlants.Rows[dgvPlants.SelectedCells[0].RowIndex].Cells["PlantID"].Value.ToString();
+                if (DialogResult.Yes == MessageBox.Show("Do you want to Delete " + name + "?", "Confirmation", MessageBoxButtons.YesNo, MessageBoxIcon.Warning))
+                {
+                    new MainForm().SqlDelete("DELETE FROM Plant WHERE PlantID = " + index);
+                }
+                new MainForm().QuerySelectAll("Plant", dgvPlants);
+            }
+            
         }
     }
 }
